Create Readme.md
I have put a basic skeleton of a Readme.md file. It still needs to be fleshed out but it gives a start. The readme is critical to get attention since it tells others what your project is about. The other file that is equally important is the license file.